Make appearance picker available in playground
