+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/MainActivity.kt
@@ -4,12 +4,10 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.stripe.android.core.version.StripeSdkVersion
+import com.stripe.android.paymentsheet.example.databinding.ActivityMainBinding
+import com.stripe.android.paymentsheet.example.playground.activity.PaymentSheetPlaygroundActivity
 import com.stripe.android.paymentsheet.example.samples.activity.LaunchPaymentSheetCompleteActivity
 import com.stripe.android.paymentsheet.example.samples.activity.LaunchPaymentSheetCustomActivity
-import com.stripe.android.paymentsheet.example.playground.activity.PaymentSheetPlaygroundActivity
-
-import com.stripe.android.paymentsheet.example.databinding.ActivityMainBinding
-import com.stripe.android.paymentsheet.example.playground.activity.AppearancePlaygroundActivity
 
 class MainActivity : AppCompatActivity() {
     private val viewBinding by lazy {
@@ -31,10 +29,6 @@ class MainActivity : AppCompatActivity() {
 
         viewBinding.launchPlaygroundButton.setOnClickListener {
             startActivity(Intent(this, PaymentSheetPlaygroundActivity::class.java))
-        }
-
-        viewBinding.appearanceButton.setOnClickListener {
-            startActivity(Intent(this, AppearancePlaygroundActivity::class.java))
         }
 
         viewBinding.version.text = StripeSdkVersion.VERSION_NAME
