+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/activity/AppearanceStore.kt
@@ -1,0 +1,14 @@
+package com.stripe.android.paymentsheet.example.playground.activity
+
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import com.stripe.android.paymentsheet.PaymentSheet
+
+internal object AppearanceStore {
+    var state by mutableStateOf(PaymentSheet.Appearance())
+
+    fun reset() {
+        state = PaymentSheet.Appearance()
+    }
+}
