+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/samples/viewmodel/PaymentSheetViewModel.kt
@@ -15,7 +15,6 @@ internal class PaymentSheetViewModel(
     val inProgress = MutableLiveData<Boolean>()
     val status = MutableLiveData<String>()
     val exampleCheckoutResponse = MutableLiveData<ExampleCheckoutResponse>()
-    val appearance = MutableLiveData(PaymentSheet.Appearance())
 
     fun statusDisplayed() {
         status.value = ""
