+++ paymentsheet-example/src/main/res/layout/activity_main.xml
@@ -65,17 +65,6 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/launch_custom_button" />
 
-        <Button
-            android:id="@+id/appearance_button"
-            android:text="@string/appearance_playground"
-            android:layout_width="0dp"
-            android:layout_height="48dp"
-            android:layout_marginEnd="16dp"
-            android:layout_marginStart="16dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/launch_playground_button" />
-
         <TextView
             android:id="@+id/version"
             android:layout_width="wrap_content"
