+++ paymentsheet-example/src/main/res/menu/menu_playground.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+    <item
+        android:id="@+id/appearance_picker"
+        android:title="Appearance"
+        android:icon="@drawable/ic_color_lens"
+        app:showAsAction="always"
+        tools:ignore="HardcodedText" />
+</menu>
